backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
pandas
openpyxl
django-cors-headers
